+++ views/groupseditdelete.php
@@ -1,24 +1,42 @@
 <?php
-include 'var.php';
-//create connection
-$connection = mysqli_connect($servername, $username, $password, $dbname);
-
-//test if connection failed
-if(mysqli_connect_errno()){
-    die("connection failed: "
-        . mysqli_connect_error()
-        . " (" . mysqli_connect_errno()
-        . ")");
+if(session_status() !== PHP_SESSION_ACTIVE) session_start();
+if(!isset($_SESSION['signed_in']) || $_SESSION['signed_in'] == false)
+{
+    echo "Illegal access";
+    echo '<a href="../index.php">Proceed to the forum CON general page</a>.';
 }
 
-$GID = (int)$_GET['GID'];
-$UID = (int)$_GET['UID'];
 
 
-echo "You have removed user with ID: $UID from group: $GID<br>";
+//connected
+else {
+
+    echo '<h6> Welcome, ' . $_SESSION['name'] . '</h6><br>';
 
 
-$query="delete from group_membership where group_id=$GID and member_id=$UID";
+//echo $_SESSION['id'];
+
+
+    include 'var.php';
+//create connection
+    $connection = mysqli_connect($servername, $username, $password, $dbname);
+
+//test if connection failed
+    if (mysqli_connect_errno()) {
+        die("connection failed: "
+            . mysqli_connect_error()
+            . " (" . mysqli_connect_errno()
+            . ")");
+    }
+
+    $GID = (int)$_GET['GID'];
+    $UID = (int)$_GET['UID'];
+
+
+    echo "You have removed user with ID: $UID from group: $GID<br>";
+
+
+    $query = "delete from group_membership where group_id=$GID and member_id=$UID";
 
 //mysqli_query($connection,$query)or die(mysqli_error());
 //if($connection->affected_rows>=1){
@@ -26,13 +44,13 @@ $query="delete from group_membership where group_id=$GID and member_id=$UID";
 //}else{
 //    echo "<p>($ud_name) Not Updated<p>";
 //}
-if (!$connection -> query($query)) {
-    echo("Error description: " . $connection -> error);
+    if (!$connection->query($query)) {
+        echo("Error description: " . $connection->error);
+    }
+
+    $connection->close();
+    echo '<td><a  href=groupsedit.php?ID=' . $GID . '>Click here to further modify the group</a></td>';
+    echo '<button onclick="self.close()">Click here to exit popup</button>';
 }
-
-$connection -> close();
-echo  '<td><a  href=groupsedit.php?ID='.$GID.'>Click here to further modify the group</a></td>';
-
 ?>
-<!input type="button" onclick="location.href='groupsedit.php';" value="Click here to further modify the group" />
-<button onclick="self.close()">Click here to exit popup</button>
+
